+++ GameViewController.m
@@ -74,7 +74,7 @@
             if(!cancelled) {
                 DefenseEvent* event = [[DefenseEvent alloc] initDefender:player action:isOutOfBounds ? PullOb : Pull];
                 if (hangtimeMilliseconds > 0) {
-                    event.pullHangtimeMilliseconds = hangtimeMilliseconds;
+                    event.pullHangtimeMilliseconds = (int)hangtimeMilliseconds;
                 }
                 [self addEvent: event];
                 [self showDidYouKnow];
